Improve segment editing
It is now possible to add, remove and change points in a permanent way.
When adding a point between two other points the time is interpolated is the difference between the two points halved.
When extending segment, adding points at the beggining or the end, the time will be interpolated using the delta time between the last two or next two points.