Change undo/redo behaviour and storage
Storing whole copies of the past state was very expensive, with three or four undos the system would colapse.
With this new method a undo function is set by the action. This function puts the state in it's original form, before the action was taken.
When redoing, the action is simply dispatched again. This way we only store n functions for n undo/redos, which are usually inexpensive.